move table view related code into extensions
